refactor(des): split HistoricoTab (584→96) em hook + componentes
Extrai verbatim de src/components/des/HistoricoTab.tsx:
- useHistoricoData: 4 queries (metas/snapshots/checkins/posição ao vivo)
  + memos derivados (cards/anosDisponiveis/cardsFiltrados/chartData/metaMedia)
  + estado de filtros
- HistoricoFiltros, FaturamentoChart, QuarterCardItem, DetalhesDialog,
  StarsRow, types.ts, format.ts (fmtBRL/fmtPct/fmtDate/quarterDates)
- +14 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
